test_roma/requirements.txt
roma-dspy
dspy-ai>=2.4.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
e2b>=0.1.0
pandas>=2.0.0
matplotlib>=3.7.0
